separate out rabbitmq setup into own file
